requirements.txt
django
djangorestframework
django-cors-headers
mysql_connector
mysql
pandas
